+++ CHANGES.rst
@@ -19,6 +19,7 @@ Minor Changes
 
 - Simplify & update internal functionality that also removes a vanishingly unlikely infinite loop condition.
 - Utilize unused package seed during package initialization.
+- Improve the derivation algorithm for the ``AffinePermutation`` 's  multiplicative key.
 - Improvements & fixes to type hints.
 - Improvements & fixes to doc strings.
 - Improvements & fixes to package documentation.